// packages/splits-sdk/src/typechain/VestingModule/VestingModule.ts
export declare namespace VestingModule {
  export type VestingStreamStruct = {
    token: PromiseOrValue<string>;
    vestingStart: PromiseOrValue<BigNumberish>;
    total: PromiseOrValue<BigNumberish>;
    released: PromiseOrValue<BigNumberish>;
  };

  export type VestingStreamStructOutput = [
    string,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    token: string;
    vestingStart: BigNumber;
    total: BigNumber;
    released: BigNumber;
  };
}